rename NoteKind to Note
