+++ src/app/shared/Note.svelte
